smtp_server: don't log load shedding rejection records
Spent some time debugging a pathological situation today with a log hook
where the result of this was like chucking a gas tanker on to a fire,
and pushed the memory utilization even higher.

The issue we're avoiding here is: when we are load shedding, we are not
in a healthy state to accept new messages. If we have a log hook
configured, logging a rejection is actually logging a new message.

Since the rejection causes simple/eager injectors to retry very quickly
it is possible to see multiples of the normal rate of message ingress
get queued up to the log hook queue, so it can very quickly make
things escalate beyond tolerable levels.