spool-util: add initial version
This isn't shipped in the packages at this time.
It's a utility for post-mortem or offline analysis
of kumomta spools.

This version assumes rocksdb, but could be tweaked to work
with a plain filesystem spool.

We could add functions in the future to migrate from one format
to another, or perform other sorts of maintenance/admin operations.