ready_queue: allow dynamically changing max_ready
This was deliberately adjusted to only change when a queue
was reapead in a prior release, in the interests of raw performance.

While that might work for most domains, it's not ideal for busy
domains that have continual traffic.

This change adopts an ArcSwap to atomically swap out the underlying
ArrayQueue.  This introduces a small amount of overhead in the hot
path to atomically acquire the current version of the queue, but
it shouldn't be enough to be noticeable in practice.