shaping: parallelize DNS resolution to improve performance
Previously we'd sequentially walk through the list of domains to resolve
their MX records.  This can be slow, especially in the context of `kumod
--validate` or when running `validate-shaping` standalone where we start
with empty MX caches.

This commit runs the lookups in parallel, with up to 128 concurrent
lookups taking place at once, so that we can get through a large
shaping configuration in much less time.