shaping: remove "Multiple domains rollup to the same site" check
This is more of a conceptual error than a hard error. We were treating
it as a validation error with the intent of getting a human to eyeball
and address it, but TSA can generate rules that will trigger this
conceptual conflict.

The conceptual conflict is this: if you have two domains with actually
conflicting definitions that rollup to the same site_name, the actual
shaping behavior will vary depending on which domain was used to
instantiate the ready queue for that site.  This was exacerbated
by the random seed used in the hashmaps used internally.

What this commit does it replace the hashmaps with orderedmaps
to preserve the user-defined ordering and remove the conflict check.

Removing the conflict check is kinda like sweeping this under the rug,
but I'm mostly comfortable with it because I think you'd need to go
reasonably out of your way to set up something conflicting, and
because it a tricky problem to detect definitive conflicts.

So, we may well come back to revisit this in the future.