amqp: adjust scheduling
The back story is that we've been trying to run down a situation
where the AMQP connections seemd to get busy, running close to 100%,
and that appears to impact some other processing.

This commit dedicates a thread per AMQP connection to manage its
connection state.  This is likely not totally necessary, but it
is desirable to move it off of the current tokio context, which
is one of our lua localset threads, in order to reduce that
contention.

I don't really like this as a solution, it's just giving us
an option to play with while we zero-in on the underlying issue.

As part of this, we now have an explicit timeout around the connect
operation.

This commit also adds a publish_with_timeout method on the amqp
connection, which combines the publish and the wait for confirmation
and puts a timeout around that operation.