lua_deliver: treat unknown errors and 421 reject as close connection
Add the ability for a lua delivery handler to use `kumo.reject(421)`
as a way to signal that it wants to close the lua delivery handler.

Treat other (non-kumo.reject) errors as something that should also
close the connection, so that misc errors will be mapped to this;
it seems simpler and more ergonomic for implementors of lua handlers
to not have to worry about the nuance of detecting and mapping the
error in the common case.  For things that need more smarts, the
user can employ pcall and match the error state explicitly.