shaping.lua: pre-match log records against shaping rules
Rather than queueing up a complete copy of all logs to send
to the TSA daemon, we now perform a match on the client side.
If the record doesn't match any rules then we won't bother
sending it to the daemon.

Otherwise, we'll queue and send.

This reduces the amount of data that we'd need to queue, thus
reducing IO pressure, and reducing inter-node bandwidth.

This new behavior is currently opt-in, but will likely
be the default by the time we cut the next release.

To opt-in, add `pre_filter = true` to the `setup_with_automation`
parameters.