logging: define new logging thread pool, migrate from localset
Simplify the logic re: threading in the logging module by adopting
the use of a dedicated Runtime for it and using that in place
of the ad-hoc current-thread runtime + localset runtime mixture
that was previously in use.

Splitting it out into a separate threadpool makes it feasible
to adjust the size of the threadpool via the pre_init event
for more tuning possibilities.