http: add native rust auth result cache
There is a measurable performance impact that can be observed
when bouncing from the http request processing flow and into
lua to validate the auth credentials, even using memoization
for the lookup.

This commit adds a small but effective cache around the result
of the lua call so that we can avoid that relatively expensive
bounce in most cases.