http: add httpinject thread pool
kumo.set_httpinject_threads can be used to tune the size.

Previously, we'd spawn the injections into the localset
pool which was not easy to tune.

Part of this change is to remove the additional extraneous
spawn around the queue insertion; we mustn't spawn that part
because it defeats the back pressure that we have from other
parts of the system.  It makes no material difference in
normal operation and helps to keep overall memory usage
bounded when the system is in distress.