kcli top: switch latency back from p90 to avg
The bucketing in the histograms on a second-to-second
time span makes the p90 less useful; it always just shows
the upper bound of one of the buckets because there's not enough
data in any bucket for the interpolation to show anything else.

We'll need to do some math over a long sampling period than the
kcli update interval for it to be more useful.

So, for now, we're showing the avg.