config: add lua_event_started counter
This is bumped each time we're about to call out to a lua event.
The idea is that you can compare the corresponding
lua_event_latency_count to get a sense of the number events that
have not yet completed and that might be blocked or otherwise
running slower than desired.