config: add gc-on-put chance
In some respects, this is conceptually redundant or similar with
the option to drop after a certain number of requests, but there
are two important differences:

* This is probabalistic rather than setting an absolute threshold,
  so it should scale with the traffic
* Rather than dropping the entire context, we just gc it, allowing
  it to be reused later.  Depending on how much garbage was collected,
  this is probably cheaper than creating a new context from scratch
  in a complex configuration.

The way this option works is that you set a percentage chance for
a gc to happen when a context is returned to the pool from 0-100.

A random number is generated and if it falls below the chance, then
a gc run is triggered.

Setting this value large can harm performance.

I've found that setting it to 1 has minimal impact while also
helping to keep the un-reaped Message handles to a bounded and
more reasonable level.

This option defaults to 0.