queue: add experimental strategy="SingletonTimerWheel"
As mentioned in a prior commit, when there are very large numbers
of scheduled queues (hundreds of thousands), the tokio timer/task
infrastructure becomes over-burdened and the system bogs down.

This commit introduces a SingletonTimerWheel strategy for the
scheduled queues.  When this strategy is in use, each scheduled
queue maintains a local HashSet of Messages rather than an
individual wheel or skip list.  The time ordering is instead
handled by a global singleton timerwheel that holds weak references
to the messages.

The definitive ownership of the Message belongs to the per-Queue
HashSet.

The global wheel ticks over once every 3 seconds for ALL queues
that have delayed messages, popping off the weak references
and attempting to upgrade them to a full Message reference.

From there, the message is resolved back to its containing queue,
and if it is found in the HashSet then it is reinserted into the
ready queue.

If either the upgrade or HashSet check fails then the message
has been either bounced or rebound and requires no further
processing.

This change makes the number of scheduled tasks for delayed
messages O(1), rather than O(number-of-queues), at the cost
of a small loss in precision of promoting the delayed messages
to the ready queue, and double the amount of Message handles;
previously it would be approx:
  number-of-delayed-messages * sizeof(pointer)
now it is 2x that amount (~ + some misc overhead)

I don't plan to add a SingletonSkipList strategy: since the list is
global it can have potentially millions of delayed messages (the entire
server's delayed mail) so we really do want O(1) insertion for this
rather than the skip lists worse insertion complexity.

Since this change eliminates the maintainer task, we need an alternative
approach for processing configuration refreshes, so this commit moves
that out of the per-queue maintainer and into a separate task that wakes
up periodically to assess all scheduled queues to see if they are due
for a refresh (according to their individual refresh_intervals) and
perform the refresh.

This change makes the config refresh a sequential operation:
only the refresh task will zip through and perform the refreshes.
Previously, you could expect to see most/all of the qmaint threads
doing this.