queue: adopt new Pruning counters for lazy scheduled metrics
This commit introduces a kumo-prometheus helper crate that
features PruningIntCounterVec and PruningIntGaugeVec types.
These vend out labelled counters of the appropriate that
that will automatically de-register from the metrics collector
when they are no longer referenced.

Previously, we had to decide in code when to do that, which was
"OK" for the Queue itself but "impossible" for the tenant/campaign
variations of those counters.

In the same theme, the ScheduledMetrics struct is now lazily allocated
when we actually have something to record against the scheduled queue.
In the through-and-through case we don't need any metrics here, so
we can skip the overhead of creating and tracking zero counters.

Likewise, the delay_due_to... counters are lazily allocated on top
of that, leading to further savings in overhead.

There's a little tweak in here that probably should be in a separate
commit to fixup reap_interval wakeups. I'm going to refactor how
reaps work in the next commit, so I'm not losing sleep over this.