queue: add bounce check to singleton wheel promotion
This is adding parity with the maintainer task which checks for
bounces prior to popping.

Here we check for each message as we promote, which is a little
bit meh, but is effectively what we do in the hot path when inserting
through to the ready queue initially; we check bounces in `insert`
prior to calling `insert_ready`.  Since we're calling insert_ready
directly, we can just make the same check explicitly for ourselves.