kcli top: adopt tokio interval with skip policy for updates
In situations where the server is under heavy stress, the metrics
endpoint can take longer than the update interval to return.
Over a period of time this can result in a backlog of pending
update requests that put more pressure on the server, and that
will also be late.

This commit switchers to an alternative interval approach that
knows to skip intervals if we're running behind.