config: remove deserialie error log
It turns out that this particular class of error is never logged,
despite the comment here about unconditionally logging, because
the default log filter excludes.

I found this because in the next commit I enable logging for
the config crate and I've found that the various checks to see
if a parameter is valid performed by the helpers will actually
trigger this log line all the time, so it has to go.