queue: fix non-singleton maintainer wakeups
I recently broke when adding the singleton maintainer task.
The nature of the issue is that the scheduled queue maintainers
are created and won't wake up to do any work for 1 day, or until
they get notified by an action like a rebind operation.

Why didn't the integration tests catch this obvious regression?
Because they didn't assert that we saw a specific number of attempts,
only that the mathmatical relationship between an arbitrary number
of records was correct.

As part of this, I found a similar sort of wakeup issue with the
skiplist queue, where adding the initial entry didn't wakeup
the queue because of the way that the Ord trait works on Option.
This commit addresses that as well.