queue: fix "priority inversion" around config_epoch
This is another case of rust's unexpectedly extended lifetimes
around lock guards leading to a bad day.

The issue was that the config_epoch would be locked when the
config refresh is processed, which can lead to a deadlock when
there are multiple other threads operating on the same queue.

The fix is to scope the lock + copy to its own function so that
the lifetime of the guard is clearly terminated by the closing
of the function scope.