egress_source: move source->site_name cache into round robin struct
Part of the pool -> source round robin needs to compute the ready
queue name (aka site_name) for each candidate source so that it
can determine whether it is suspended.  This resolution was previously
called on every call to next().

At the higher level, in the Queue logic for deciding which ready
queue should be picked, we had a cache for this information, and
it was worthwhile to do that in order to shave off some small
overheads during the hot-path enqueue.

This commit pushes that cache down into the round robin state
so that we can avoid that overhead more often.

This is a micro-optimization, so don't expect a massive uplift
from this change.