fix: minor issue applying the wrong value to a tokio runtime builder
I wanted to play with these low level runtime parameters and I noticed
that, due to copypasta, we ended up setting the max_io_events_per_tick
to the same value we used for event_interval.

I couldn't detect any noticeable difference with this "broken" vs
"fixed" on my test system, of even with the values adjusted to
extremes, so I think the impact of this boo-boo is very minor.

While I'm in here looking at this, apply the same values to the "main"
tokio runtime well.