throttle: add force_local flag
I want to use a local-only throttle in an experiment, so this commit
prepares the throttle crate for that.

You can now use `local:200/hr` to request a local-only throttle,
even when redis throttles have been enabled.

While I'm in here, since this changes the debug repr and there
are a number of snapshot tests that are affected by this change,
I've added Debug and Display impls that match the canonical
textual format.