dns-resolver: randomize the list of hosts at a given pref level
Previously, we would deterministically produce the list of hosts
and use the same thing for each session that we established outbound.
This could result in biasing the outgoing traffic onto a single host
and result in lower overall throughput.

This commit shuffles the set of hosts at a given preference level
to probabalistically distribute the load among them.