ready_queue: fix ready_queues not aging out until memory shortage
The config epoch changes broke the periodic idling out of the ready
queue maintainer tasks, which could lead to increase memory utilization
over time, both in terms of the ReadyQueue related structures, and
clogging up the counter registry.

This commit adds a timer that is updated after each maintainer wakeup.