ready_queue: fix ready_queues not aging out until memory shortage II
Following on from the change in
b584d472b125391ccd26b203c358ecbc5fc6243b, it is possible to get stuck
looping if we don't update the next due time.

This could do with more scrutiny overall, so I'm not calling this done,
but this commit doesn't look like it can make things worse, and should
help things in the meantime for folks that are brave enough to be
running `main` in a more production facing role.