shaping: provider: matching could panic with non-ascii input
There is something funky happening the bigger picture, as I wouldn't
expect to see exotic accented characters in the domains that are
being fed through the system.

Regardless, there is a string manipulation issue in our suffix
matching function where we trip over this input:

```
thread 'spoolin-6' panicked at crates/kumo-api-types/src/shaping.rs:804:31:
byte index 2 is not a char boundary; it is inside 'é' (bytes 1..3) of `réputation.net`
```

We're only doing slicing stuff here so that we can apply the case
insensitive comparison function efficiently.

This commit resolves this issue by using the built-in ends_with function
instead, which makes the matcher not pass for mixed case inputs.

In practice, I think this is fine, because the input domain string
is a product of the dns-resolver crate which normalizes names to
lower case.

I want to avoid an explicit conversion to lower case for this because
that uses heap memory and introduces overheads in a relatively
hot code path.